requirements.txt
pytesseract>=0.3.10
PyMuPDF>=1.20.0
pandas>=1.5.0
openpyxl>=3.0.0
Pillow>=9.0.0
numpy>=1.21.0
